pyproject.toml
[project]
name = "semantic-cache"
version = "0.1.0"
description = "Semantic caching web service using Redis and sentence-transformers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "redis>=5.2.0",
    "redisvl>=0.1.0",
    "sentence-transformers>=3.0.0",
    "numpy>=2.0.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "ty>=0.0.12",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["semantic_cache"]

[tool.ty]
# Ty configuration (Astral's fast type checker)
environment."python-version" = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
